Created a new folder for JavaScript and added the files for Simon Game project
The Simon game is a very popular memory game where the player is supposed to remember the order in which the coloured tiles flash and replicate it in his/her move.
*It is a single-player game.

This project's driver code has been coded in JavaScript and the frontend interface is built using HTML and CSS.

Each time, an additional tile gets added to the order, making it difficult for the player to remember.

The moment the player fails to replicate the exact order or presses a wrong tile, the game is over.