Created a new Folder for Python and added Tic Tac Toe project
The game is played on a grid that's 3 squares by 3 squares.

You are X, your friend is O. Players take turns putting their marks in empty squares.

The first player to get 3 of her marks in a row (up, down, across, or diagonally) is the winner.

When all 9 squares are full, the game is over.