Create PerfectNumber.java
This is a Java Program to Check if a given Number is Perfect Number. A perfect number is a positive integer that is equal to the sum of its proper positive divisors excluding the number itself.
Enter any integer as an input. Now we use modulo operation along with loops and if conditions to check given number is a perfect number or not.