Fibonacci series using java.
This program will print Fibonacci series.
Fibonacci series is a series of numbers. The pattern will print number which is the sum of the preceding two numbers.
for example if we input 5 then the output would be:
0 1 1 2 3    (i.e. we will get fibonacci series upto 5 numbers)


CODE EXPLANATION:
for example if we enter 5 as input then it gets store in "n".
-->a=0
-->b=1
-->a (that is 0) will be printed.
entering while loop until count <= n-1(4 in this case):
-counter will get b's value (i.e counter = 1)
-b = b + a (which is 1)
-now a will get b's value as b's value is stored in counter (refer line 14)
-count++ means now count's value will get increased by 1.
-finally a's value will be printed.

NOTE: while loop will run until count's value is equal to n-1.

SO FINAL OUTPUT will be : 0 1 1 2 3


This file will be helpful to others as many questions in platforms such as hackerrank , hackerearth , etc are related to Fibonacci series.